Changing connectToState function nonsense.
