+++ index.js
@@ -1,1 +1,1 @@
-export * from "./src/";
+export { CreditCardInput } from "./src/CreditCardInput.js";
