+++ src/LiteCreditCardInput.js
@@ -5,12 +5,13 @@ import {
   StyleSheet,
   Image,
   TouchableOpacity,
+  LayoutAnimation,
   findNodeHandle
 } from "react-native";
 
 import Icons from "./Icons";
 import CCInput from "./CCInput";
-import { InjectedProps } from "./connectToState";
+import { SharedProps } from "./CreditCardInput";
 import _ from 'lodash';
 
 const INFINITE_WIDTH = 1000;
@@ -18,7 +19,7 @@ const INFINITE_WIDTH = 1000;
 /* eslint react/prop-types: 0 */ // https://github.com/yannickcr/eslint-plugin-react/issues/106
 export default class LiteCreditCardInput extends Component {
   static propTypes = {
-    ...InjectedProps,
+    ...SharedProps,
 
     placeholders: PropTypes.object,
     inputStyle: Text.propTypes.style,
@@ -46,48 +47,43 @@ export default class LiteCreditCardInput extends Component {
             result : result.concat(key);
     }, []);
 
-    console.error('CCDEBUG: props changed: ' + JSON.stringify(diffProps));
+    console.log('CCDEBUG: props changed: ' + JSON.stringify(diffProps));
     let diffState = _.reduce(nextState, (result, value, key) => {
         return _.isEqual(value, this.state[key]) ?
             result : result.concat(key);
     }, []);
 
-    console.error('CCDEBUG: state changed: ' + JSON.stringify(diffState));
+    console.log('CCDEBUG: state changed: ' + JSON.stringify(diffState));
     return diffProps.length > 0 || diffState.length > 0;
   }
 
   componentWillReceiveProps (newProps) {
     if (this.props.focused !== newProps.focused) {
-      console.error('CCDEBUG: props changed new focus, call focus');
+      console.log('CCDEBUG: props changed new focus, call focus');
       this._focus(newProps.focused);
     }
-
-    let diff = _.reduce(newProps, (result, value, key) => {
-        return _.isEqual(value, this.props[key]) ?
-            result : result.concat(key);
-    }, []);
-    console.error('CCDEBUG: props changed: ' + JSON.stringify(diff));
   }
 
   _focusNumber() {
-    console.error('CCDEBUG focus number');
+    console.log('CCDEBUG focus number');
     this._focus("number");
   }
 
   _focusExpiry() {
-    console.error('CCDEBUG focus expiry');
+    console.log('CCDEBUG focus expiry');
     this._focus("expiry");
   };
 
   _focus(field) {
-    console.error('CCDEBUG lite input focus: ' + field);
+    console.log('CCDEBUG lite input focus: ' + field);
 
     if (!field || this.props.focused == field) {
       return;
     }
 
-    console.error('CCDEBUG lite input focus: ' + this[field]);
+    console.log('CCDEBUG lite input focus: ' + this[field]);
     this[field].focus();
+    LayoutAnimation.easeInEaseOut();
   }
 
   _handleSubmit() {
@@ -133,9 +129,9 @@ export default class LiteCreditCardInput extends Component {
   render() {
     let { focused, values: { number }, inputStyle, status: { number: numberStatus } } = this.props;
     let last4Value = numberStatus == "valid" ? number.substr(number.length - 4, 4) : "";
-    console.error('CCDEBUG: destructured focused: ' + focused);
+    console.log('CCDEBUG: destructured focused: ' + focused);
 
-    console.error('CCDEBUG: props focused: ' + this.props.focused);
+    console.log('CCDEBUG: props focused: ' + this.props.focused);
     let showRightPart = !focused || (focused != "number");
     let creditNumberStyle = [];
     let infoStyle = [];
@@ -145,34 +141,32 @@ export default class LiteCreditCardInput extends Component {
       creditNumberStyle = [s.hiddenViewStyle, s.leftPart];
       infoStyle = [s.expandedViewStyle, s.rightPart];
       onPressFunc = this._focusNumber.bind(this);
-      console.error('CCDEBUG: showRightPart true');
+      console.log('CCDEBUG: showRightPart true');
     } else {
-      console.error('CCDEBUG: showRightPart false');
-      creditNumberStyle = [s.expandedViewStyle, s.leftPart];
+      console.log('CCDEBUG: showRightPart false');
+      creditNumberStyle = [s.expandedViewStyle, s.leftPart, {marginLeft: 20}];
       infoStyle = [s.hiddenViewStyle, s.rightPart];
       onPressFunc = this._focusExpiry.bind(this);
     }
 
-    console.error('CCDEBUG number style: ' + JSON.stringify(StyleSheet.flatten(creditNumberStyle)));
-    console.error('CCDEBUG info style: ' + JSON.stringify(StyleSheet.flatten(infoStyle)));
+    console.log('CCDEBUG number style: ' + JSON.stringify(StyleSheet.flatten(creditNumberStyle)));
+    console.log('CCDEBUG info style: ' + JSON.stringify(StyleSheet.flatten(infoStyle)));
     return (
       <View style={s.container}>
-        <View style={creditNumberStyle}
-              key={'leftSide'}>
-          <CCInput {...this._inputProps("number")}
-              containerStyle={s.numberInput} />
-        </View>
+        <CCInput {...this._inputProps("number")}
+            containerStyle={[s.numberInput, creditNumberStyle]}
+            key={'leftSide'}/>
         <TouchableOpacity onPress={onPressFunc}>
           <Image style={s.icon}
                  source={{ uri: Icons[this._iconToShow()] }} />
         </TouchableOpacity>
         <View style={infoStyle}
               key={'rightSide'}>
-          <TouchableOpacity onPress={this._focusNumber}
+          <TouchableOpacity onPress={this._focusNumber.bind(this)}
                             style={s.last4}>
             <CCInput field="last4"
                      value={last4Value}
-                     onFocus={this._focusNumber}
+                     onFocus={this._focusNumber.bind(this)}
                      inputStyle={[s.input, inputStyle]}
                      containerStyle={[s.last4Input]} />
           </TouchableOpacity>
@@ -218,7 +212,6 @@ const s = StyleSheet.create({
   },
   numberInput: {
     width: INFINITE_WIDTH,
-    marginLeft: 20,
   },
   expiryInput: {
     width: 60,
